feat: token validation trade service

+++ src/main/java/com/project/stocker/service/TradeService.java
@@ -10,10 +10,12 @@ import com.project.stocker.entity.Orders;
 import com.project.stocker.entity.Stock;
 import com.project.stocker.entity.Trade;
 import com.project.stocker.entity.User;
+import com.project.stocker.jwt.JwtUtil;
 import com.project.stocker.repository.OrdersRepository;
 import com.project.stocker.repository.StockRepository;
 import com.project.stocker.repository.TradeRepository;
 import com.project.stocker.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -31,23 +33,31 @@ public class TradeService {
     private final StockRepository stockRepository;
     private final UserRepository userRepository;
     private final OrdersRepository ordersRepository;
+    private final JwtUtil jwtUtil;
 
     @Autowired
     private TradePublisher tradePublisher;
 
     //sell order publish
-    public TradeCreateResponseDto sellOrders(TradeCreateRequestDto ordersCreatRequestDto) {
+    public TradeCreateResponseDto sellOrders(TradeCreateRequestDto ordersCreatRequestDto, HttpServletRequest request) {
+        String token = jwtUtil.getJwtFromRequest(request);
+        ordersCreatRequestDto.setToken(token);
         tradePublisher.publishSellOrders(ordersCreatRequestDto);
         return new TradeCreateResponseDto(HttpStatus.OK.value(), "매도 주문 처리 중");
     }
+
+
 
     //sell order subscriber
     public TradeCreateResponseDto subSellOrders(TradeCreateRequestDto ordersCreateRequestDto) {
         String stockName = ordersCreateRequestDto.getStock();
         Long quantity = ordersCreateRequestDto.getQuantity();
         Long buyPrice = ordersCreateRequestDto.getPrice();
+        String token = ordersCreateRequestDto.getToken();
 
-        User user1 = userRepository.findById(1L).orElseThrow(() ->
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user1 = userRepository.findByEmail(email).orElseThrow(() ->
                 new IllegalArgumentException("id가 1인 유저가 존재하지 않습니다."));
 
         Stock stock = (Stock) stockRepository.findByCompany(stockName).orElseThrow(() ->
@@ -63,15 +73,25 @@ public class TradeService {
     }
 
     //sell orders update
-    public TradeUpdateResponseDto sellUpdate(TradeUpdateRequestDto sellUpdateDto) {
+    public TradeUpdateResponseDto sellUpdate(TradeUpdateRequestDto sellUpdateDto, HttpServletRequest request) {
 
         Long tradeId = sellUpdateDto.getTrade_id();
         Long quantity = sellUpdateDto.getQuantity();
         Long buyPrice = sellUpdateDto.getPrice();
 
+        String token = jwtUtil.getJwtFromRequest(request);
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("email이 없습니다"));
+
         // tradeId validation
         Orders orders = (Orders) ordersRepository.findById(tradeId).orElseThrow(() ->
                 new IllegalArgumentException("해당 매도 신청이 존재하지 않습니다."));
+
+        if (!orders.getBuyer().equals(user) && !orders.getSeller().equals(user)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
 
         // trade status validation
         orders.setQuantity(quantity);
@@ -83,13 +103,23 @@ public class TradeService {
     }
 
     //sell orders delete
-    public TradeDeleteResponseDto sellDelete(TradeDeleteRequestDto sellDeleteDto) {
+    public TradeDeleteResponseDto sellDelete(TradeDeleteRequestDto sellDeleteDto, HttpServletRequest request) {
 
         Long tradeId = sellDeleteDto.getTrade_id();
+
+        String token = jwtUtil.getJwtFromRequest(request);
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("email이 없습니다"));
 
         // trade Id validation
         Orders orders = (Orders) ordersRepository.findById(tradeId).orElseThrow(() ->
                 new IllegalArgumentException("해당 매수 신청이 존재하지 않습니다."));
+
+        if (!orders.getBuyer().equals(user) && !orders.getSeller().equals(user)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
 
         //user validation
         ordersRepository.delete(orders);
@@ -98,7 +128,9 @@ public class TradeService {
     }
 
     //buy orders publish
-    public TradeCreateResponseDto buyOrders(TradeCreateRequestDto ordersCreatRequestDto) {
+    public TradeCreateResponseDto buyOrders(TradeCreateRequestDto ordersCreatRequestDto, HttpServletRequest request) {
+        String token = jwtUtil.getJwtFromRequest(request);
+        ordersCreatRequestDto.setToken(token);
         tradePublisher.publishBuyOrders(ordersCreatRequestDto);
         return new TradeCreateResponseDto(HttpStatus.OK.value(), "매수 주문 처리 중");
     }
@@ -108,15 +140,18 @@ public class TradeService {
         String stockName = ordersCreateRequestDto.getStock();
         Long quantity = ordersCreateRequestDto.getQuantity();
         Long buyPrice = ordersCreateRequestDto.getPrice();
+        String token = ordersCreateRequestDto.getToken();
 
-        User user2 = userRepository.findById(2L).orElseThrow(() ->
-                new IllegalArgumentException("id가 2인 유저가 존재하지 않습니다."));
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("유저가 존재하지 않습니다."));
 
         Stock stock = (Stock) stockRepository.findByCompany(stockName).orElseThrow(() ->
                 new IllegalArgumentException("해당 종목이 존재하지 않습니다."));
 
         Orders trade = new Orders.Builder(quantity, buyPrice, stock)
-                .buyer(user2)
+                .buyer(user)
                 .build();
 
         ordersRepository.save(trade);
@@ -125,15 +160,25 @@ public class TradeService {
     }
 
     //buy orders update
-    public TradeUpdateResponseDto buyUpdate(TradeUpdateRequestDto buyUpdateDto) {
+    public TradeUpdateResponseDto buyUpdate(TradeUpdateRequestDto buyUpdateDto, HttpServletRequest request) {
 
         Long tradeId = buyUpdateDto.getTrade_id();
         Long quantity = buyUpdateDto.getQuantity();
         Long buyPrice = buyUpdateDto.getPrice();
 
+        String token = jwtUtil.getJwtFromRequest(request);
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("email이 없습니다"));
+
         // tradeId validation
         Orders orders = (Orders) ordersRepository.findById(tradeId).orElseThrow(() ->
                 new IllegalArgumentException("해당 매수 신청이 존재하지 않습니다."));
+
+        if (!orders.getBuyer().equals(user) && !orders.getSeller().equals(user)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
 
         // trade status validation
         orders.setQuantity(quantity);
@@ -145,13 +190,23 @@ public class TradeService {
     }
 
     //buy orders delete
-    public TradeDeleteResponseDto buyDelete(TradeDeleteRequestDto buyDeleteDto) {
+    public TradeDeleteResponseDto buyDelete(TradeDeleteRequestDto buyDeleteDto, HttpServletRequest request) {
 
         Long tradeId = buyDeleteDto.getTrade_id();
+
+        String token = jwtUtil.getJwtFromRequest(request);
+        String email = jwtUtil.getUserEmailFromToken(token);
+
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+                new IllegalArgumentException("email이 없습니다"));
 
         // trade Id validation
         Orders orders = (Orders) ordersRepository.findById(tradeId).orElseThrow(() ->
                 new IllegalArgumentException("해당 매수 신청이 존재하지 않습니다."));
+
+        if (!orders.getBuyer().equals(user) && !orders.getSeller().equals(user)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
 
         //user validation
         ordersRepository.delete(orders);
